submitter tests work!!

+++ tests/test_qtools_submitter.py
@@ -11,21 +11,28 @@ from subprocess import PIPE
 
 import tests
 
+HOSTNAME = subprocess.Popen('HOSTNAME', stdout=subprocess.PIPE).communicate()[
+    0].strip()
+
+# Global variable to test if we're on a server, e.g. TSCC or oolite (
+# "compute" means one of the oolite compute nodes)
+ON_SERVER = set([HOSTNAME]) & set(['tscc', 'oolite', 'compute'])
+
 
 class Test(unittest.TestCase):
     def test_main(self):
         commands = ['date', 'echo testing PBS']
         job_name = 'test_qtools_submitter'
-        submit_sh = '%s/%s.sh' % (tests.get_test_dir(), job_name)
+        submit_sh = '%s.sh' % (job_name)
         sub = Submitter(queue_type='PBS', sh_file=submit_sh,
                         command_list=commands,
                         job_name=job_name)
-        job_id = sub.write_sh(submit=True, nodes=1, ppn=16,
-                                 queue='home-yeo', walltime='0:01:00')
-        true_result_string = '''#!/bin/sh
+        job_id = sub.write_sh(submit=False, nodes=1, ppn=16,
+                              queue='home-yeo', walltime='0:01:00')
+        true_result_string = '''#!/bin/bash
 #PBS -N test_qtools_submitter
-#PBS -o %s/test_qtools_submitter.sh.out
-#PBS -e %s/test_qtools_submitter.sh.err
+#PBS -o test_qtools_submitter.sh.out
+#PBS -e test_qtools_submitter.sh.err
 #PBS -V
 #PBS -l walltime=0:01:00
 #PBS -l nodes=1:ppn=16
@@ -36,7 +43,7 @@ class Test(unittest.TestCase):
 cd $PBS_O_WORKDIR
 date
 echo testing PBS
-''' % (tests.get_test_dir(), tests.get_test_dir())
+'''
         true_result = true_result_string.split('\n')
 
         # with open(submit_sh) as f:
@@ -47,23 +54,25 @@ echo testing PBS
             self.assertEqual(true.strip().split(), test.strip().split())
 
         # Make sure the job ID is a single (potentially multi-digit) integer
-        self.assertRegexpMatches(job_id, '^\d+$')
-        subprocess.Popen(["qdel", job_id],
-                                 stdout=PIPE)
+        # But only do this if we're on TSCC or oolite
+        if ON_SERVER:
+            self.assertRegexpMatches(job_id, '^\d+$')
+            subprocess.Popen(["qdel", job_id],
+                             stdout=PIPE)
 
     def test_wait_for_pbs(self):
         commands = ['date', 'echo testing PBS']
         job_name = 'test_qtools_submitter_wait_for_pbs'
-        submit_sh = '%s/%s.sh' % (tests.get_test_dir(), job_name)
+        submit_sh = '%s.sh' % (job_name)
         sub = Submitter(queue_type='PBS', sh_file= submit_sh,
                         command_list=commands,
                         job_name=job_name, wait_for=['11111'])
-        job_id = sub.write_sh(submit=True, nodes=1, ppn=16,
-                                 queue='home-yeo', walltime='0:01:00')
-        true_result_string = '''#!/bin/sh
+        job_id = sub.write_sh(submit=False, nodes=1, ppn=16,
+                              queue='home-yeo', walltime='0:01:00')
+        true_result_string = '''#!/bin/bash
 #PBS -N test_qtools_submitter_wait_for_pbs
-#PBS -o %s/test_qtools_submitter_wait_for_pbs.sh.out
-#PBS -e %s/test_qtools_submitter_wait_for_pbs.sh.err
+#PBS -o test_qtools_submitter_wait_for_pbs.sh.out
+#PBS -e test_qtools_submitter_wait_for_pbs.sh.err
 #PBS -V
 #PBS -l walltime=0:01:00
 #PBS -l nodes=1:ppn=16
@@ -75,7 +84,7 @@ echo testing PBS
 cd $PBS_O_WORKDIR
 date
 echo testing PBS
-''' % (tests.get_test_dir(), tests.get_test_dir())
+'''
         true_result = true_result_string.split('\n')
 
         # with open(submit_sh) as f:
@@ -86,23 +95,23 @@ echo testing PBS
             self.assertEqual(true.strip().split(), test.strip().split())
 
         # Make sure the job ID is a single (potentially multi-digit) integer
-        self.assertRegexpMatches(job_id, '^\d+$')
-        subprocess.Popen(["qdel", job_id],
-                                 stdout=PIPE)
+        if ON_SERVER:
+            self.assertRegexpMatches(job_id, '^\d+$')
+            subprocess.Popen(["qdel", job_id], stdout=PIPE)
 
     def test_wait_for_array_pbs(self):
         commands = ['date', 'echo testing PBS']
         job_name = 'test_qtools_submitter_wait_for_pbs'
-        submit_sh = '%s/%s.sh' % (tests.get_test_dir(), job_name)
+        submit_sh = '%s.sh' % (job_name)
         sub = Submitter(queue_type='PBS', sh_file= submit_sh,
                         command_list=commands,
                         job_name=job_name, wait_for_array=['11111'])
-        job_id = sub.write_sh(submit=True, nodes=1, ppn=16,
-                                 queue='home-yeo', walltime='0:01:00')
-        true_result_string = '''#!/bin/sh
+        job_id = sub.write_sh(submit=False, nodes=1, ppn=16,
+                              queue='home-yeo', walltime='0:01:00')
+        true_result_string = '''#!/bin/bash
 #PBS -N test_qtools_submitter_wait_for_pbs
-#PBS -o %s/test_qtools_submitter_wait_for_pbs.sh.out
-#PBS -e %s/test_qtools_submitter_wait_for_pbs.sh.err
+#PBS -o test_qtools_submitter_wait_for_pbs.sh.out
+#PBS -e test_qtools_submitter_wait_for_pbs.sh.err
 #PBS -V
 #PBS -l walltime=0:01:00
 #PBS -l nodes=1:ppn=16
@@ -114,7 +123,7 @@ echo testing PBS
 cd $PBS_O_WORKDIR
 date
 echo testing PBS
-''' % (tests.get_test_dir(), tests.get_test_dir())
+'''
         true_result = true_result_string.split('\n')
 
         # with open(submit_sh) as f:
@@ -125,10 +134,11 @@ echo testing PBS
             self.assertEqual(true.strip().split(), test.strip().split())
 
         # Make sure the job ID is a single (potentially multi-digit) integer
-        self.assertRegexpMatches(job_id, '^\d+$')
-        subprocess.Popen(["qdel", job_id],
-                                 stdout=PIPE)
+        if ON_SERVER:
+            self.assertRegexpMatches(job_id, '^\d+$')
+            subprocess.Popen(["qdel", job_id], stdout=PIPE)
+
 
 if __name__ == "__main__":
-    #import sys;sys.argv = ['', 'Test.test_main']
+    # import sys;sys.argv = ['', 'Test.test_main']
     unittest.main()
